// packages/codegen/tests/fixtures/planet.d.ts
/**
 * Generated by @ts-pf/codegen. Do not edit.
 * catalogVersion: 1
 * catalogHash: sha256:5b7371c3adad6facdb936e1c7d0453f303cd31630a5b7b3efe82dc74b5481b25
 */

import type { ContractProcedure } from '@ts-pf/contract'

type Phantom<T> = {
  readonly '~standard': {
    readonly version: 1
    readonly vendor: 'ts-pf-codegen'
    readonly types: { readonly input: T; readonly output: T }
    readonly validate: (value: unknown) => { readonly value: T }
  }
}

export type Contract = {
  planet: {
    find: ContractProcedure<
      { id: number },
      { id: number; name: string },
      {
        NOT_FOUND: {
          status: 404
          data: Phantom<{ id: number }>
        }
      }
    >
    list: ContractProcedure<void, { id: number; name: string }[]>
    chat: ContractProcedure<{ prompt: string }, AsyncIterable<{ token: string }>>
  }
}